Added the ability to upload a files. Model, repository, service, view.

+++ src/main/java/ru/dreamjob/service/SimpleVacancyService.java
@@ -1,6 +1,7 @@
 package ru.dreamjob.service;
 
 import org.springframework.stereotype.Service;
+import ru.dreamjob.dto.FileDto;
 import ru.dreamjob.model.Vacancy;
 import ru.dreamjob.repository.VacancyRepository;
 
@@ -12,23 +13,45 @@ public class SimpleVacancyService implements VacancyService {
 
     private final VacancyRepository repository;
 
-    public SimpleVacancyService(VacancyRepository repository) {
+    private final FileService fileService;
+
+    public SimpleVacancyService(VacancyRepository repository, FileService fileService) {
         this.repository = repository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Vacancy save(Vacancy vacancy) {
+    public Vacancy save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
         return repository.save(vacancy);
+    }
+
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        var file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
+        var vacancyOptional = findById(id);
+        if (vacancyOptional.isEmpty()) {
+            return false;
+        }
+        fileService.deleteById(vacancyOptional.get().getFileId());
         return repository.deleteById(id);
     }
 
     @Override
-    public boolean update(Vacancy vacancy) {
-        return repository.update(vacancy);
+    public boolean update(Vacancy vacancy, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return repository.update(vacancy);
+        }
+        var oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        var isUpdated = repository.update(vacancy);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
